backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.0
pydantic>=2.0.0
python-multipart>=0.0.5
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
requests>=2.26.0
web3>=5.24.0
httpx>=0.23.0
idna>=2.8
pyjwt>=2.0.0
python-dotenv>=0.19.0 